pyproject.toml
[project]
name = "7-ai-code-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate==1.10.0",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "authlib==1.6.4",
    "backoff==2.2.1",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography==45.0.6",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docx2txt==0.9",
    "effdet==0.4.1",
    "emoji==2.14.1",
    "et-xmlfile==2.0.0",
    "executing==2.2.0",
    "faiss-cpu==1.12.0",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "flask==3.1.1",
    "flatbuffers==25.2.10",
    "fonttools==4.59.0",
    "fpdf2==2.8.4",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "google-cloud-core==2.4.3",
    "google-cloud-firestore==2.21.0",
    "google-cloud-vision==3.10.2",
    "googleapis-common-protos==1.70.0",
    "gpt4all==2.8.2",
    "greenlet==3.2.4",
    "grpcio==1.74.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "hf-xet==1.1.9",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.4",
    "humanfriendly==10.0",
    "idna==3.10",
    "instructorembedding==1.0.1",
    "ipykernel==6.30.1",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.7",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyterlab-widgets==3.0.15",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.74",
    "langchain-google-genai==2.1.9",
    "langchain-openai==0.3.29",
    "langchain-text-splitters==0.3.9",
    "langdetect==1.0.9",
    "langsmith==0.4.13",
    "lxml==6.0.0",
    "markdown==3.8.2",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.5",
    "matplotlib-inline==0.1.7",
    "motor==3.7.1",
    "mpmath==1.3.0",
    "msoffcrypto-tool==5.4.2",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "narwhals==2.0.1",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "nltk==3.9.1",
    "numpy==2.2.6",
    "olefile==0.47",
    "omegaconf==2.3.0",
    "onnx==1.18.0",
    "onnxruntime==1.22.1",
    "openai==1.99.5",
    "opencv-python==4.12.0.88",
    "openpyxl==3.1.5",
    "orjson==3.11.1",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pandas==2.3.1",
    "parso==0.8.4",
    "pdf2image==1.17.0",
    "pdfminer-six==20250506",
    "pi-heif==1.1.0",
    "pikepdf==9.10.2",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycocotools==2.0.10",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pymongo==4.14.1",
    "pypandoc==1.15",
    "pyparsing==3.2.3",
    "pypdf==5.9.0",
    "pypdf2==3.0.1",
    "pypdfium2==4.30.0",
    "pyreadline3==3.5.4",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "python-iso639==2025.2.18",
    "python-magic==0.4.27",
    "python-multipart==0.0.20",
    "python-oxmsg==0.0.2",
    "python-pptx==1.0.2",
    "pytz==2025.2",
    "pywin32==311",
    "pyyaml==6.0.2",
    "pyzmq==27.0.1",
    "rapidfuzz==3.13.0",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.27.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.1",
    "selenium==4.35.0",
    "sentence-transformers==5.1.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "sqlalchemy==2.0.42",
    "stack-data==0.6.3",
    "streamlit==1.48.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.11.0",
    "timm==1.0.19",
    "tokenizers==0.21.4",
    "toml==0.10.2",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.55.0",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "unstructured==0.18.11",
    "unstructured-client==0.42.2",
    "unstructured-inference==1.0.5",
    "unstructured-pytesseract==0.3.15",
    "urllib3==2.5.0",
    "validators==0.35.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "weaviate-client==4.16.10",
    "webdriver-manager==4.0.2",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "werkzeug==3.1.3",
    "widgetsnbextension==4.0.14",
    "wrapt==1.17.3",
    "wsproto==1.2.0",
    "xlrd==2.0.2",
    "xlsxwriter==3.2.5",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
